Add query params for /rbac/current, /rbac/users, rbac/forms.  Allow caller to filter on permissions and roles
