+++ app/src/forms/rbac/controller.js
@@ -43,7 +43,7 @@ module.exports = {
   },
   getCurrentUser:  async (req, res, next) => {
     try {
-      const response = await service.getCurrentUser(req.kauth.grant.access_token);
+      const response = await service.getCurrentUser(req.kauth.grant.access_token, req.query);
       res.status(200).json(response);
     } catch (error) {
       next(error);
